chore(security): neutralize react-native-executorch telemetry and defer init
